Add dummy security-group rule to avoid SSH EC2 connections
At the moment it is not possible to avoid the creation of the default
security group by ElasticBeanstalk. The group adds a rule that opens
the SSH port to the outside world. We don't want that, so a workaround
was to add a dummy rule that only allows connections from our security
group (which in practice is the same machine, so no outside
connections are allowed).

According to the ElasticBeanstalk roadmap they will add an option to
disable this default security group soon:
Issue 44 in repo https://github.com/aws/elastic-beanstalk-roadmap