Assume the SSL certificate will be already available at create time
If we already have a SSL certificate configured we can configure the
ELB listeners and rules properly.

Creating the certificate within the templates it's not possible
because the creation would be stuck until the validation is
completed. That can be quick if we have control over the DNS server,
but sometimes they belong to external parties and that could slow the
process.

As a workaround we will use a self-signed certificate. That
certificate can't be created or imported from Cloudformation, so we
will do it in a previous step using the API.