Add elasticbeanstalk sample project with custom hooks
We will use the hooks in order to download the EB environment
variables from the SSM Parameter store.
- The 'hooks/prebuild' hooks will take care of installing the Babashka and
our SSM utility tool for using it in the next steps.
- The 'hooks/predeploy' hook will be executed when a new version of
the application is being deployed. The script download the variables
from SSM and concats them to the '.env' file.
- The 'confighooks/predeploy' hook is executed when the configuration
is changed in EB. We need to do it here too not to lose the changes we
did when deploying the app version.