Add additional SSM policy for Full read permissions
* TODO: create a custom policy to only add the permissions
required. AmazonSSMManagedInstanceCore seems to include both
permissions for SSM Parameter Store and SSM Session Manager. But it
doesn't include a crucial permission that we need which is
GetParameterByPath. So the AmazonSSMReadOnlyAccess give all the read,
list and describe permissions. Therefore, we should come up with a new
policy with a more fine grained permissions setting.