Update settings.edn for handling docker deploym.
[Re #2]

Now we handle in a different way docker-compose related deployment options, in
order to let the user choose between `external` (already managed docker
images) and `container` options (launched with the rest of the containers).

The main point here is to be sure we don't allow different docker-compose
deployment options for different environments related to deployments (such as
test and prod), as that would increase a lot the complexity and we do not
mainly need this feature. Thus, we only allow different options for
development and the non-development (for deployment) groups.

We have only reflected this change in the Keycloak related configuration, but
we need to spread it to the rest of the configuration and also handle it in
the rest of the code.