Extract DB version out of envs and hardcode Docker host
[Re #2]

We have extracted DB version config out of environments-related config, since
it is not worth it to make this config so flexible.

Besides, we have also removed the persistence-related docker host name config,
since it is also not worth it to provide such flexiblity, taking into account
the complexity it adds.

Fixed DB config tag name in settings.edn.