Update settings.edn with Grafana profile data
[Re #2]

Updated settings file to reflect a complete configuration for Grafana, using
same approach as for Keycloak.

In this case, we have not included yet any extra configuration related to
Keycloak, since we have not yet a clear strategy to provide inter-profile
dependencies.

Beware that we still need to deal with the changes needed in the code to
manage the Grafana profile from settings.edn file.