Connect Cognito, S3, RDS settings file with AWS Cloudformation
[Re #2]
When we have a multiple or single choice group we need to check the
'value' to see if a specific keyword is in the list. But when we are
working with Mustache templates it's not possible to check if an array
contains something. So, to facilitate that an make other use-cases
simpler decided to always add a 'enabled true' key/value to the
selected choices.