Wait if AWS rate is exceeded when putting parameters to SSM
[Re #2]
In other projects we have been using Diheard's rate-limiters, but that
library is not compatible with BB (because it's a wrapper over a Java
library).

As we the documentation doesn't clearly which is the quota for using
the API we decided to retry the requests when they fail after waiting
for a few seconds. We run the function with different waiting periods
to find the most appropiate one, but the results are quite
incosistent. We decided to leave it in 3 seconds.