Cognito profile: add compose file to inject profile related env vars
* Some profiles' services are cloud based like Cognito and at some
point they might need to add environment variables to the docker
environment of other services. The way we do this is by creating a
compose file defining the additional environment variables to the
docker service the external service is affecting.

For example in this case, if we choose Cognito as our authentication
service we need to include three additional environment variables to
the docker compose environment of the application. We do that by
creating a compose file in the Cognito profile that simply add those
variables to the app service environment.