Refactor profile execution to use multimethods
[Re #2]
In order to avoid re-evaluating the registry map when a change is made
to the code we decided to use multimethods. So now evaluating the
profile namespace will be enough.

We also decided to have multiple hooks where the profiles can execute
code. That will allow us to implement the prod/test SQL generation.

Even if we use the multimethod we need to keep a collection of
profiles to determine the order in which they have to be executed.