Fix hikari-cp failing at key initialization
[Re #2]
The `duct/database.sql.hikaricp` library doesn't support specifying
the db connection details individually and it throws an exception at
startup. We are going to make a PR to the library, but while that
happens we decided to have an alternative fork.

Also decided to get rid of 'duct./module.sql' because we were not
really using it. We were just taking advantage of the dependencies
that it adds, but not using the code itself.