Support memory to be configurable in containers
[Re #2]

Added configuration support about max memory allowed for each container
(MB), for each environment.

We have added this support for each service's docker container (if that
deployment type is chosen).

In the case of the app and nginx containers, we have included a explicit Core
profile in `settings.edn` file that should be always pre-selected to include
this configuration of memory from the user, for each enviroment.