Fix keycloak routes and environment variables
[Re #2]

* Keycloak URI is different depending on the deployment type. If it's
container there are two different URIs:
 - External -> keycloak frontend application URI
 - Internal -> keycloak docker container.
In the container deployment type case both External and Internal URIs
are needed. External would be used by the FE application to redirect
the user to the Keycloak web interface and to do calls to the server
from the client. And the Internal one would be used by the
application's BE to retrieve the JWKS from the Keycloak container
service, in order to authenticate ID tokens issued by
Keycloak.

* Fixed the config-edn routes, which should use the api-routes key instead of
routes.