Fix cloudformation templates upload
[Re #2]

* When the application is packed as an uberjar we cannot access the
JAR file contents that easily with io/resource. So, our approach is
the same as we did for the profiles' bootstrap files. We are unzipping
the JAR file to a temporary folder on the user's machine and using the
files from. Once, the operation finishes the temporary folder is
deleted.

* This is not the idea solution or final. Ideally we should do this
once for the bootstrap process as we need this in two places as of
now. This could probably happen later on with other features. So we
need to think about a final solution for this problem in the long term.