Only delete project specific volumes when running start-dev.sh
[Re #2]
The scripts in HOP should only deal with the volumes created by HOP
itself. Deleting other volumes might be unexpected for the end users,
and could cause accidental data loss.

Currently we use a single Docker container, which is the PostgreSQL
one. Postgres defines a volume in the Dockerfile, so it's
automatically created when the container is started. By default the
automatically created volume doesn't have any identifier, so it was
impossible for us to identify and delete the volume explicitly. By
defining the volume explicitly the volumes get the
'com.docker.compose.project' that we can use in the deletion.