Support AWS Region to be configurable
[Re #2]

We have added a new config entry in `settings.edn` file for letting the user configure its prefered AWS Region to create the project.
For this, we have needed to use a function for creating the AWS Client with the AWS Region to be provided in runtime.

In order to work with an up-to-date version of AWS Regions available, we have updated S3 and AWS Endpoint-related dependencies as well.

During this change, we have also unified the way we pass arguments to the AWS API, so we use a single map of arguments in most cases, where no other separation is needed.

Added additional print statements to show AWS Region information when showing dev-profile related info.

As we might have different bucket names for storing Cloudformation
templates, we have needed to add the AWS Region as part of the bucket
name, so we can have different buckets per region, so we can use
different stack names per region.