Implement create-settings-file sub-command as alias to copy-settings
Add the create-settings-file sub-command as an alias to the existing
copy-settings command (so both keep working, the video tutorial video
is still valid, etc). But the documentation should only mention the
"create-settings-file" command.

"copy-settings" is how the operation it is currently implemented,
while "create-settings-file" is what the end user operation is about.

[Re: #5]