Add failed stack events to error-details output when it fails
[Re #6]

* If CFN stack creation fails we want to know why it did directly from
the CLI output. For that we are using the CFN DescribeStackEvents API
to retrieve the stack events. Since we are only checking this in the
case of failure, we filter the events and format them to output as
`error-details`. Note that we are checking for the ResourceType for
the resource that failed, because it can be a stack. If it's a stack,
we need to describe that stack events as well to finally find the real
root cause of the failure.