Add saving to file button in settings editor UI
[Re #9]

Now we allow the user to save the current modified settings.edn file
using a button that when is clicked it creates a link on the fly that
is clicked to download a blob with the settings.edn file, downloaded
automatically by the browser. The idea is that this saved file can be
used later to create a new project.