Use readlink instead of realpath
realpath is not available in macOS/OSX. That means
compose-project-name.sh outputs an error when run in macOS/OSX.

That script was also missing the set -eu -o pipefail line that we add
to all of our bash scripts. That is the reason why the script didn't
exit when the above error was produced. And we get something that
worked in many cases (but not all).

[Closes: #11]