Update desc of HOP CLI commands using Docs repo
We have updated the commands description to match the ones from the
official documentation (`hop-documentation` repo). Now they are more
verbose but we keep consistent descriptions that are more descriptive.
We are also using the terms of the defined glossary for the updated
docs.
There is some exception in the update, as the HOP CLI's
documentation for the `bootstrap create-settings-file` command were
better than what we had in the docs, so we have kept that doc instead.