Improve docker-env-vars.sh logic for checking docker-compose version
* The previous logic wasn't enough as newer versions of docker-compose
v1.x outputs a different return code when providing a flag argument
that doesn't exist. For example docker-compose version 1.27.x will
return 1 code when executing 'docker-compose config
--no-normalize' (that flag doesn't existing on versions 1.x) but
version 1.29.x will return a 0 code. So instead of relying on that,
now we parse the of CLI and check if it's 1.x or something else.