Make SQL instructions post-install step a bit clearer.
Right now the SQL script instructions follow the sentence about
executing the script itself, whithout any separation. And it's a bit
unclear where the SQL commands to execute begin and end. Add a bit of
whitespace before and after, and add some SQL comments saying where it
starts and ends.