Impl. on-premises profile and deployment-target
[Re #4]

* In this profile we are not implementing deployment per se but providing
the necessary files to deploy the final artifact in a GNU/Linux
machine.
* Including CI scripts and app bundle creation scripts to help during
deployment.
* SSL Termination. There are two options, NGINX or HTTPS Portal[1]. The
former is just a NGINX configuration file that the end-user can use to
configure SSL termination on its own NGINX reverse proxy. The latter
is an automated way of adding SSL termination using HTTPS Portal
Docker container.
* The first time installation process will be redacted in the offical
HOP documentation with all the steps needed to run and maintain the
application.

[1] - https://github.com/SteveLTN/https-portal