Fix permissions in the deployment bundle files
Certain CI environments set an overly permissive umask. That means the
git checkout and any files we create in the CI scripts have too many
permissions. And when we copy the existing checkout files into the
bundle, we also copy the permissive access. We need to copy the files
with their original permissions, because some of then need to be
executables, etc. But we don't want the "extra" permissions added
because of the CI environments umask.

So we take a two pronged approach:

- We set a reasonable umask at the beginning of the bundle creation
  script.
- When creating the final bundle tar file, we remove the most dangerous
  ones ('group' and 'other' write permissions).