Correctly manage environment entries with '$' chars in them
If the value of one environment entry (of a docker-compose service)
contains a '$' character in it, we previously assumed that it
contained a variable name in it, and we didn't set it to null.

But it may happen that even if the value contains a '$', it doesn't
contain a variable name! The docker-compose specification allows to
use the '$$' combination to actually specify a '$' character in the
value.

We just need to use a tighter regular expression to detect variables
names. The docker-compose specification allows the `$VARIABLE` and
`${VARIABLE}` syntax, where the variable name can only have certain
characters in them. Use that criteria to detect variable names.