"reset" volume paths that are supposed to be relative
We shouldn't turn them into absolute paths (much less with an
arbitrary path). We need to keep them as relative paths, like they
initially were. That allows us to place the on-premises deployment
files in arbitrary places. And docker-compose will still be able to
use the files/directories that are located in the bundle itself (as
opposed to those that are elsewhere, like the persistent data
directories).