Local development environment doesn't use AWS KMS keys
The local development environment stores its environment variables in
a local file (.env), and not in AWS Parameter Store. So there is no
need to create an AWS KMS key at all (as we only use those keys to
encrypt/decrypt the values of the environment variables in the AWS
Parameter Store, which is not used in this case).

[Re: #18]