Create KMS key aliases with the names specified by the user
We have explicit settings for the aliases of the AWS KMS keys, in the
settings file. But the values specified there were not used when
creating the AWS KMS keys and setting their aliases. We
programmatically built the aliases values by concatenating the project
name and the environment name (dev, test, prod).

This resulted in errors when parts of the bootstrapping code tried to
use the alias names specified in the settings file, as those aliases
might not exist in KMS (in fact, they wouldn't, unless the values
configured by the user matched the values programmatically generated
by our CloudFormation templates).

So make sure we the values provided by the user in the settings file.

[Re: #18]