Enable the use of PKCE for OIDC auth
OAuth 2.0 public clients (and by extension, OpenID Connect clients, as
it's based on it) utilizing the Authorization Code Grant are
susceptible to the authorization code interception attack. The PKCE
specification[1] describes the attack as well as a technique to
mitigate against the threat through the use of Proof Key for Code
Exchange (PKCE, pronounced "pixy").

As the SPA we use for the front-end is a public client that is using
the Authorization Code Grant, we want to enable the use of PKCE for
this authentication flow against the IdP (Keycloak in our case).

[1] https://www.rfc-editor.org/rfc/rfc7636