Use at least ten chars for the SHA1 prefix for commits
The git SHA1 prefix that we obtain when running git log --oneline
locally in our repos, and the SHA1 prefixes that are calculated by
ci/common-vars.sh (to build what we internally call the "version tag")
in all the projects that are HOP based, sometimes don't "match". In
the sense that, the same SHA1 prefix computed by each of them, for the
same commit, has a different length. The output of one of those
commands is sometimes longer than the other (e.g., one outputs a 7
characters SHA1 prefix, and the other a 8 characters SHA1 prefix).

This can introduce confusion for us, and let us use the wrong SHA1
prefix in the Docker images names that we use to deploy things, if
done manually (this is less of an issue in projects configured to be
deployed to AWS Elastic Beanstalk, as the deployment bundles are
created in the CI/CD pipeline as well, using the same SHA1 prefix in
call cases).

This is not a theoretical issue, as some of use have suffered it one
or more times 😁

So we have decided that we are going to use at least 10 chars for the
SHA1 prefixes. This assures us that we will have a fixed (and
predictable) amount of characters in the "version tags", no matter
where we run that command. The documentation got git rev-parse says
that the number we specify there is the minimum amount of chars in the
SHA1 prefix. There may be more, if there are more than one commit in
the repo with the same prefix. Using 10 characters makes that
unlikely.

But if we hit that unlikely case, we would always get the same amount
of characters for that commit (until some point in the future when a
third potential commit would collide with that amount of chars; but
that wouldn't be a big problem, as we would have already deployed the
version with the "colliding tag" in the past, etc.)

[Fixes: #23]