Ignore everything in .clj-kondo and .lsp dirs except their config
And the .gitignore file that is used to ignore the rest of the things,
of course ;-)

The reason is that several third-party libraries include their own
clj-kondo and lsp configuration, and they "pollute" their directories
with files that we don't want to include in our commits by
mistake. And we don't want to play "whac-a-mole" with each new library
that adds things there.