Use a new way of merging docker-compose files
Our previous approach used `yq` with a lot of difficult to understand
and maintain DSL. It also had a very annoying shortcoming: it
couldn't deal with array entries that had values with environment
variables, and entries that didn't.

We have finally bitten the bullet and decided to use Babashka to
implement the docker-compose files merging and avoid both issues
altogether.