Set default memory limits for containers
While we always define the MEMORY_LIMIT_XXXX environment variables in
the local development .env files, and we also define them in AWS SSM
Parameter Store (for AWS deployments), there is one edge-case where
the docker-compose files are used and the variables are still not
defined. Which makes docker-compose error out.

This use case happens when we deploy to AWS using
ElasticBeanstalk. Because of size limits imposed by ElasticBeanstalk,
we don't store them in the ElasticBeanstalk application definition. We
store them in Parameter Store. And when the application is being
deployed or restarted, we have platform hooks that run some code to
get the variables from Parameter Store and inject them in the .env
file that ElasticBeanstalk will use to run the application.

The platform hooks run at two different moments of the application
deployment and start up. We retrieved and injected the variables in
the second hook, and that worked just fine in the past. But apparently
ElasticBeanstalk has introduced some additional actions in between
those two different moments, and that something tries to use the
docker-compose files before we can even attempt to injected the
variables in the .env file.

Environment variables that are defined as such in the `environment:`
stanzas of the docker-compose files are fine, because even if they are
not defined yet, they are simple assumed to be `null` (which is a
valid value, from the syntax point of view). But container memory
limit syntax doesn't allow `null`, and triggers a syntax error. Which
makes that intermediate step to fail, which makes ElasticBeanstalk to
stop the deployment or restart.

Setting some default memory limits avoids that issue, because now we
will have valid syntax values. Those default limits won't actually be
used when the application is run. Because by that time, the second
hook would have been run, and the right values would have been
retrieved from Parameter Store and injected in the .env file.