Restart Keycloak if it fails to start
Sometimes we have a race condition between the startup of Postgres and
the startup of Keycloak. Because we use a very simple healtcheck for
Postgres, docker-compose may think the Postgres service is up and
running when it may not completely true.

The fact is that Postgres restarts several times during the initial
startup (after creating the db the first time, after installing the
extensions, etc.). The first healthcheck may find Postgres healthy,
and thus launch Keycloak. But by the time Keycloak tries to connect to
the database to create its tables, it may be restarting and thus
unavailable. This makes Keycloak startup fail, and the container
remains stopped.

By adding the "restart: unless-stopped" setting, we tell
docker-compose to restart the Keycloak container every time it stops,
unless manually stopped by us.