Random password generation sometimes uses incompatible characters
When we fixed issue #24, we focused on the AWS ElasticBeanstalk use
case (where we populate the .env file from SSM Parameter Store). But
it turns out we missed the other big use case where the same problem
may happen: when creating the local development .env file (or the test
/ production .env files for on-premises deployments).

This commit addresses that second use case. There is a bit of code
churn in this commit because we wanted to handle both use cases using
the same utility functions (so we apply the quoting policy in a single
place). That in turn meant that we had to refactor a bunch of SSM
Parameter Store related code, because in that case we directly used
the data in the shape that Parameter Store returned. And that was
incompatible with the data shape used in the rest of the code dealing
with docker-compose environment variables.