Fix on-premises-files copying
Because we didn't specify a `:dst` property for those files, the files
were being copied at the root of the new project directory. Which
didn't make any sense at all.

Now we explicitly say that we want those files copied inside the
`on-premises-files` directory, at the top of the new project
directory.