Make "load-frontend-app" template rendering a bit more flexible.
There are certain profiles (e.g., Cognito and Keycloak profiles) that
need to inject their own code in the templates used by the "frontend"
profile. To do that, we used the "project.load-frontend-app" mustache
dictionary.

The problem is that depending on the profile, they may need to inject
that code in different namespaces. And we had an dictionary with a
single level of keys to render the different code pieces (requires,
extra events to dispatch on app initialization, extra code to add
events, subscriptions or other re-frame artifacts to the main client
namespace, etc). Which meant that the same code pieces would be
injected in all the namespace templates where we used that dictionary.

We have added an additional top level to that dictionary, to be able
to specify which namespace is supposed to be used with the (now second
level) dictionary that holds the actual values, when rendering the
templates.