Enhance error handling when missing sub-commands
If the user types a partial command (e.g., `hop aws cognito` instead
of `hop aws cognito get-id-token`), we didn't handle the exception
thrown by Babashka CLI about the missing sub-command(s). Same if we
type the name of the sub-command wrong.

Now we handle those situations and let the user know about the kind of
problem we encountered.