Use Launch Templates when creating ElasticBeanstalk environments
AWS ElasticBeanstalk has phased out launch configurations and
replaced them with launch templates. But, by default,
ElasticBeanstalk still uses launch configurations[1]. The only way,
for the time being, to force ElasticBeanstalk to use launch templates
is by specifying some OptionSettings with "special" values (see [1]
again for details).

We have decided to use the `RootVolumeType` option as the way to
achieve it, as the least invasive change.

[1] See details, especially for phase out dates, at
https://docs.aws.amazon.com/elasticbeanstalk/latest/dg/environments-cfg-autoscaling-launch-templates.html

[Re: #27]