Refactor <-pgvalue for JSON(B) values
The initial implementation mostly copied the example code from the
next.jdbc documentation. That code included attaching meta-data (the
original Postgresql type for the value) to the value decoded from
JSON(B). Probably to round-trip from the database to the application
and back to the database, and maintain the same type without tracking
it manually in the app.

But `with-meta` only works on values that implement
`clojure.lang.IObj`. And we sometimes get JSON(B) values that are
simple strings, booleans or numbers, which don't implement IObj. And
the call to `with-meta` throws an Exception.

Given that we don't use the original Postgresql type at all, and thus
don't use the meta-data at all, we are removing it. That we we can
work with values that don't implement clojure.lang.IObj too.

Once we are refactoring the function, we alto take the opportunity to
explicitly set the separator used to build keywords. It turns out that
we routinely have JSON(B) keys that contain digits in them, instead of
just alphabetic characters (.e.g., "p023m345_a123c"). And
`->kebab-case-keyword` treats digits as "word separators" while
building the keywords. That is, "p023m345_a123c" gets converted to
`:p-023m-345-a-123c`, instead of `:p023m345-a123c`, which is what we
expect. Setting the separator to the underscore character fixed the
issue.