Add support for custom CA certificates
When getting ID tokens from Keycloak we sometimes need to use custom
CA certificates (e.g., for local development environments).

Instead of completely skipping the SSL certificate verification step
and accept any SSL certificate, we want to be able to verify as valid
only those certificates issued by one of a specific set of CAs.

This new option allows us to provide a path to a "certificate store"
file. That file should container, in PEM format, the certificated of
the CAs we want to use for verification.