Fix --temporary? option handling
Commit a314a3a55d9 broke the handling of this value. We (wrongly)
assumed that the coercing function would get a value that would either
be true (if specified), false (also, if specified), or nil (if not
specified). But in the two former cases we get "true" and "false", as
strings. Thus, using the standard `boolean` function always returned a
truthy value (they are non-nil strings), no matter what was specified
in the command line.

Using the Babashka CLI provided `:boolean` coercer does the right
thing in those cases.