Allow Cognito to automatically send messages to verify and confirm
[Re: #28]

By default, Cognito doesn't send account confirmation emails (or SMS
messages) any longer. The feature has to be explicitly enabled in the
"Sign Up" section (when using the AWS Console). When disabled (the
default), Cognito expects the applications to use the Admin API to
obtain those codes.

The sending of the confirmatino messages has to be enabled
irrespective of whether you want Cognito to send emails using their
default email or AWS SES.

The Cloudformation documentation for AWS::Cognito::UserPool doesn't
seem to mention any parameter that can control this configuration
setting. Fortunately, from the question
https://stackoverflow.com/questions/75939787 (How to enable Cognito's
"Attribute verification and user account confirmation" from
cloudformation)

    I've been reading back and forth AWS documentations, but I am not
    able to find how to proper set the settings shown in the image
    below using cloudformation template. What I want to achieve is
    that, when a user signs up, Cognito sends an automatic email with
    a custom template presenting the user with the code to confirm its
    account.

    [...]

    I thought the keys VerificationMessageTemplate and
    EmailConfiguration would do the job, but no email is received.

There is an answer at https://stackoverflow.com/a/75940061 that says:

    I think the key name (AutoVerifiedAttributes) for this feature is
    quite misleading since it may suggest that the attributes listed are
    auto-verified once a user has signed up.

    Therefore, adding:

          AutoVerifiedAttributes:
            - email

    does the job.

So this is what we are doing.