Allow Cognito to automatically send messages to verify and confirm
[Re: #28]

By default, Cognito doesn't send account confirmation emails (or SMS
messages) any longer. The feature has to be explicitly enabled in the
"Sign Up" section (when using the AWS Console). When disabled (the
default), Cognito expects the applications to use the Admin API to
obtain those codes.

The sending of the confirmatino messages has to be enabled
irrespective of whether you want Cognito to send emails using their
default email or AWS SES.

There is a related configuration ("Keep original attribute values when
an update is pending") that is not enabled by default when you create
a Cognito User Pool via Cloudformation.

That configuration is controlled by the Cloudformation parameter named
AttributesRequireVerificationBeforeUpdate. Its documentation states:

    Requires that your user verifies their email address, phone
    number, or both before Amazon Cognito updates the value of that
    attribute. When you update a user attribute that has this option
    activated, Amazon Cognito sends a verification message to the new
    phone number or email address. Amazon Cognito doesn’t change the
    value of the attribute until your user responds to the
    verification message and confirms the new value.

    When AttributesRequireVerificationBeforeUpdate is false, your user
    pool doesn't require that your users verify attribute changes
    before Amazon Cognito updates them. In a user pool where
    AttributesRequireVerificationBeforeUpdate is false, API operations
    that change attribute values can immediately update a user’s email
    or phone_number attribute.

Given that we are configuring mandatory confirmations for email
changes, it makes sense to enable this other parameter by default
too. This way users can still sign in with the original email address
before they have confirmed the code to verify the new value. And they
can only they can only sign in with the new attribute value after they
have confirmed the code to verify the new value.

Which prevents the users from getting locked out of their accounts if
they change their emails but never receive the email to confirm the
change.