Rework Cloudformation syntax for extra Cognito permissions
Commit 9599872972 added extra permissions to set or reset the Cognito
users' password for the local development role only (and only for the
local development User Pool)

The syntax used in that commit, will still theoretically valid, is no
longer accepted by later versions of the Cloudformation linter than
the one we used when we introduced that change. The net result is that
the linter refuses to validate the template and CI/CD fails.

While not as compact (and redundancy free) as the old syntax, using
Fn::FindInMap intrinsic function allows us to achive the same effect.