Run lein clean with enough permissions to do its job
By bypassing the default entry point for the application container,
and using the bash shell as the entry point, we ensure that the
command is run by the root user (instead of the hop user).

The reason why we need to do this is that developers sometimes
manually run some commands inside the application container by
executing `docker exec -ti container-name some-command and-args`. The
problem with that is that, when using `docker exec`, the command is
run as the root user inside the container (unless we use the `--user`
option of `docker exec` to specify another user).

That in turn means that if that command creates any files or
directories, or makes copies of existing files or directories, all of
those files and directories will be owned by the root user.

If any of those files are inside the directories where our
tooling (lein, clj-kondo, etc) is expected to run, we will have
problems.

Because all of our tooling, when run inside the container, runs as the
hop user, which is a regular user. And when run from the host, it also
runs as a regular use (that of the developer, which has the same
uid/gid of the hop inside the container).

So if the tooling (either inside the container, or on the host) has to
delete, move or otherwise change any of those files or directories
owned by root in any way, it will fail in most[1] cases. Because it
won't have the required permissions to perform the operation.

By running the `lein clean` command as the root user, we avoid all of
those problems.

[1] There are a few special cases where it will succeed. E.g., if a
file is owned by root, but it's inside a directory owned by hop/the
developer user, and the tooling simply wants to delete the
file. Because to delete a file we don't need to have any permission on
the file itself, but on the directory that contains the file.