[Core] Fix docker-env-vars.sh yq approach in latest version
The previous strategy is broken in the newest version of yq, producing
unexpected results. We have changed the approach to follow the
strategy of a specific example we found in the documentation[1] that
does what we want, update multiples items in a map. This approach
works correctly with the 4.45.2 version of yq and we get the expected
result from "docker-env-vars.sh"

[1] - https://github.com/mikefarah/yq/blob/v4.45.2/pkg/yqlib/doc/operators/select.md#select-multiple-items-in-a-map-and-update

[Re: #32], [Closes: #32]