Remove awyeah-api debug log entries when getting an id token
awyeah-api produces debug log entries to let us know what the source
of the AWS credentials it is going to use for the current request
is. But when getting an id token, those debug log entries end up in
the standard output. And they get mixed with the output that we really
want to get, especially if we are asking for the id token in raw
format. Because most of the time, when asking for the raw format, we
are going to consume that output from another tool).

From now on, we should only get `warn` or higher log levels in the
output.