Fix regular expression that isn't valid in Javascript
ClojureScript regular expressions are transpiled to Javascript regular
expressions, with the `/v` modifier (to be able to use Unicode
characters in them). In Javascript, when using the `/v` modificer, the
hyphen character (`-`) becomes a bit special when used inside a
character class (like we did in the problematic regex). We cannot
longer use it "naked", even it it's positioned at the beginning or the
end of the character class (although this works when not using the
`/v` modificer). We need to quote it with a backslash.

That backslash quoting also works without the `/v` modifier in
ClojureScript, and it works in Clojure (in the JVM) too.